Time: 36 ms (94.65%), Space: 59.5 MB (39.66%) - LeetHub

+++ 1744-number-of-ways-to-form-a-target-string-given-a-dictionary/1744-number-of-ways-to-form-a-target-string-given-a-dictionary.cpp
@@ -12,7 +12,7 @@ public:
         const int mod = 1000000007;
         vector<vector<int>>dp(n+5,vector<int>(m+5,-1));
 
-        auto add = [&](long long a,long long b)->int{
+        auto add = [&](int a,int b)->int{
             a+=mod;
             b+=mod;
             return ((a%mod)+(b%mod))%mod;
